Hacky fix for UR ee_link bug
The ee_link transform should be fixed to the wrist_3_link, but it seems to be published in the robot base frame when connected to the real robot, which results in a different orientation than during simulation. This adds another frame that is untouched by the bug, but it doesn't solve the underlying problem. To be investigated.